included first version YAI
YAI is included "forcefully". a div's innerHTML is set to the HTML of the Vue file,
and scripts are included with manual script tags. scripts could be moved to start of
application so that they're already loaded by the time YAI is displayed, and CSS
has to be reviewed (graph CSS died in the process, apparently; probably conflicting
'app' ids. beautiful work as a proof of concept, probably need a more automated
way of converting existing applications.